pyproject.toml
[project]
name = "doc-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "alembic==1.13.1",
    "fastapi==0.109.2",
    "langchain==1.2.7",
    "langchain-core==1.2.7",
    "langchain-text-splitters==1.1.0",
    "mysql-connector-python==8.3.0",
    "openpyxl==3.1.2",
    "pinecone-client==3.1.0",
    "pydantic[email]==2.12.5",
    "pydantic-settings==2.12.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "sqlalchemy==2.0.27",
    "uvicorn[standard]==0.27.1",
    "xlrd==2.0.1",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "sentence-transformers==2.3.1",
    "torch>=2.2.0",
    "groq==1.0.0",
    "hf-xet>=1.2.0",
    "slowapi>=0.1.9",
    "requests>=2.32.5",
]
